+++ site-cookbooks/jenkins/recipes/job.rb
@@ -7,22 +7,40 @@
 # All rights reserved - Do Not Redistribute
 #
 
-node[:jenkins][:jobs].each do |job|
-  ruby_block "update job - #{job}" do
-    block do
-      xml = config(node.chef_environment, job)
-      begin
-        client.post("/job/#{job}/config.xml", IO.read(xml), basic_auth.merge(crumb))
-      rescue Net::HTTPServerException => e
-        if e.message == '404 "Not Found"'
-          body = IO.read(config(node.chef_environment, job))
-          header = {'Content-Type' => 'text/xml'}.merge(basic_auth).merge(crumb)
-          client.post("/createItem?name=#{job}", body, header)
-        else
-          raise e
-        end
-      end
-    end
+node[:jenkins][:polling_jobs].each do |job_name|
+  template 'tmp/config.xml' do
+    source "jobs/#{node.chef_environment.sub('compute', 'production')}/polling.xml.erb"
+    owner 'root'
+    group 'root'
+    mode '0755'
+    variables :app_info => node[job_name]
+  end
+
+  upsert_job do
+    job_name job_name
+    file_name 'tmp/config.xml'
+  end
+end
+
+node[:jenkins][:deploy_jobs].each do |job_name|
+  template 'tmp/config.xml' do
+    source 'jobs/deploy.xml.erb'
+    owner 'root'
+    group 'root'
+    mode '0755'
+    variables :app_info => node[job_name.split('-').first]
+  end
+
+  upsert_job do
+    job_name job_name
+    file_name 'tmp/config.xml'
+  end
+end
+
+node[:jenkins][:other_jobs].each do |job_name|
+  upsert_job do
+    job_name job_name
+    file_name config(node.chef_environment, job_name)
   end
 end